feat: restructure and start install process
